Added 6-size.c

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long long int b;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long int)sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long int)sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long int)sizeof(a));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long int)sizeof(b));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long int)sizeof(f));
+	printf("Size of a char: %d byte(s)\n", (int)sizeof(c));
+	printf("Size of an int: %d byte(s)\n", (int)sizeof(i));
+	printf("Size of a long int: %d byte(s)\n", (int)sizeof(a));
+	printf("Size of a long long int: %d byte(s)\n", (int)sizeof(b));
+	printf("Size of a float: %d byte(s)\n", (int)sizeof(f));
 	return (0);
 }
